Move error logic to error service, remove helper
