+++ src/services/error.ts
@@ -1,4 +1,5 @@
-import { CELL_ERROR_HIGHLIGHT_COLOR, DEBUG } from "../constants";
+import { ADMIN_EMAIL, CELL_ERROR_HIGHLIGHT_COLOR, DEBUG } from "../constants";
+import { sendEmail } from "./email";
 
 // Highlights a cell and adds a note to it to notify of an error
 export function highlightProcessingError(
@@ -15,4 +16,14 @@ export function highlightProcessingError(
     .setNote(message);
 
   SpreadsheetApp.getUi().alert(`An error occurred during processing. ${message}`);
+}
+
+// A wrapper to catch errors and notify an administrative email about issues
+export function wrapper(callback: () => void) {
+  try {
+    callback()
+  } catch (e) {
+    sendEmail("Error in trigger", e.message, ADMIN_EMAIL)
+    throw new Error(e.message)
+  }
 }
